Minor change to Usage (add --execute flag)
This is a super minor suggested change to the usage documentation. Adding the `--execute` flag to nbconvert will first execute the notebook and then export it to PDF.

I ran into a case where I had saved a notebook without the cell outputs and of course the PDF file also had no outputs.